% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{A_to_igraph}
\alias{A_to_igraph}
\title{Convert adjacency matrix to taxonomy tree}
\usage{
A_to_igraph(A)
}
\arguments{
\item{A}{phylogeny inspired adjacency matrix}
}
\description{
Convert adjacency matrix to taxonomy tree
}
