% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_taxa_table}
\alias{sample_taxa_table}
\title{Sample taxa table used for simulating the microbiome data}
\format{
A data frame of a sample taxa table with following information
\describe{
  \item{Class}{data_frame of a sample taxa table used in the simulation study}
  \item{Family}{data_frame of a sample taxa table used in the simulation study}
  \item{Genus}{data_frame of a sample taxa table used in the simulation study}
  \item{Kingdom}{data_frame of a sample taxa table used in the simulation study}
  \item{Order}{data_frame of a sample taxa table used in the simulation study}
  \item{Phylum}{data_frame of a sample taxa table used in the simulation study}
  \item{unique}{data_frame of a sample taxa table used in the simulation study}
  }
Mishra et al. (2023) \emph{Tree aggregated factor regression model}
}
\usage{
data(sample_taxa_table)
}
\description{
Sample taxa table used for simulating the microbiome data
}
\keyword{datasets}
