% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simulate_data_mbiome}
\alias{simulate_data_mbiome}
\title{Simulate multivariate response matrix using covariates, adjacency matrix,
svd decomposition of the coefficient matrix (U,D,V), sample size,
intercept, signal to noise ratio and covariance matrix parameter of multivariate
error.}
\usage{
simulate_data_mbiome(X, A, U, D, V, n, C0, snr, rho = 0.5)
}
\arguments{
\item{X}{log transformed microbial abundance data}

\item{A}{phylogeny inspired adjacency matrix}

\item{U}{left singular vector of the svd decomposition of the coefficient matrix}

\item{D}{singular values of the svd decomposition of the coefficient matrix}

\item{V}{right singular vector of the svd decomposition of the coefficient matrix}

\item{n}{sample size}

\item{C0}{intercept value}

\item{snr}{signal-to-noise ratio}

\item{rho}{error matrix correlation}
}
\description{
Simulate multivariate response matrix using covariates, adjacency matrix,
svd decomposition of the coefficient matrix (U,D,V), sample size,
intercept, signal to noise ratio and covariance matrix parameter of multivariate
error.
}
\author{
Aditya Mishra \email{akmishra@mdanderson.org}
}
