% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taro.R
\name{taro_control}
\alias{taro_control}
\title{Set of parameters that allows us to control the execution of the TARO algorithms.}
\usage{
taro_control(
  mu = 1,
  nu = 1.1,
  MMerr = 0.001,
  MMiter = 100,
  outTol = 1e-08,
  outMaxIter = 500,
  inMaxIter = 200,
  inTol = 1e-04,
  lamMaxFac = 1,
  lamMinFac = 1e-10,
  gamma0 = 2,
  elnetAlpha = 0.95,
  alpha0 = 1,
  se1 = 1,
  spU = 0.25,
  spV = 0.25
)
}
\arguments{
\item{mu}{penalty parameter used in enforcing orthogonality}

\item{nu}{penalty parameter used in enforcing orthogonality (incremental rate of mu)}

\item{MMerr}{tolerance in the majorization maximization(MM) algorithm for computing initial values when missing value occurs}

\item{MMiter}{maximum number iterations in the MM algorithm}

\item{outTol}{tolerance of convergence of outer loop in CURE}

\item{outMaxIter}{maximum number of outer loop iteration in CURE}

\item{inMaxIter}{maximum number of inner loop iteration in CURE}

\item{inTol}{tolerance value required for convergence of inner loop in CURE}

\item{lamMaxFac}{a multiplier of calculated lambda_max}

\item{lamMinFac}{a multiplier of determining lambda_min as a fraction of lambda_max}

\item{gamma0}{power parameter in the adaptive weights}

\item{elnetAlpha}{elastic net penalty parameter}

\item{alpha0}{parameter allows us to promote aggregation in the tree guided regression}

\item{se1}{set the parameter for promoting 1se rule (se1 = 1) in cross-validation}

\item{spU}{maximum proportion of nonzero elements in each column of U}

\item{spV}{maximum proportion of nonzero elements in each column of V}
}
\value{
a list of controlling parameter.
}
\description{
list of parameters for controlling taro fitting
}
\examples{
#require(taro)

control <- taro_control()

}
\references{
Mishra et al. (2023) \emph{Tree aggregated factor regression model}
}
