% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taro.R
\name{taro_sim}
\alias{taro_sim}
\title{Simulation model}
\usage{
taro_sim(
  n = 300,
  q = 50,
  nrank = 3,
  rho = 0.5,
  snr = 0.25,
  intercept = 0.5,
  taxa_table = NULL,
  setting = 1,
  sparsity = 0.05
)
}
\arguments{
\item{n}{sample size of the simulated data}

\item{q}{number of continuous multivariate response}

\item{nrank}{number of simulated latent factor (rank of the coefficient matrix)}

\item{rho}{parameter defining correlated error}

\item{snr}{signal to noise ratio}

\item{intercept}{intercept of the true multivariate model}

\item{taxa_table}{input taxa table for the simulated microbiome data}

\item{setting}{choice of the effect size of the coefficient along the nodes of the tree.}

\item{sparsity}{desired percentage of the non-zero entries
Setting 1: primarily selects rare genera to have significant effect;
Setting 2: primarily selects most abundant genera to have significant effect
Setting 3: primarily selects leaf nodes to have significant effect
Setting 4: primarily selects nodes in the phylogentic tree to have significant effect}
}
\value{
return(list(X=X, Y =Y, A = A, U = U, D=D, V=V, C0 = C0))
  \item{Y}{Generated response matrix}
  \item{X}{Generated predictor matrix}
  \item{A}{Taxonomy inspired adjacency matrix}
  \item{tree}{Pass reference phylogenetic tree for visualization}
  \item{U}{Left singular vectors of the low-rank coefficient matrix}
  \item{V}{Right singular vectors of the low-rank coefficient matrix}
  \item{D}{Singular values of the low-rank coefficient matrix}
  \item{C0}{Simulated intercept of the coefficient matrix}
}
\description{
genertate simulated data for a tree aggregated factor regression model
}
\examples{
#require(taro)

# Simulate data from a sparse factor regression model
snr <- .25; xrho <- 0.5; nrank <- 3; q <- 50; n = 300; intercept = 0.5
rho = 0.5; sparsity = 0.05; setting = 2;  # error correlation
\donttest{
sim.sample <- taro_sim(n, q, nrank, rho, snr, intercept,
                        taxa_table = NULL, setting, sparsity)
Y <- sim.sample$Y
X <- sim.sample$X
}
}
\references{
Mishra et al. (2023) \emph{Tree aggregated factor regression model}
}
